New branch
[HotsBpHelper] 忽略训练库

[HotsBpHelper] 优化自动模式下的手动体验

[HotsBpHelper] 将HeroSelectorVm以列表形式放入一个HeroSelectorWindowVm中,优化HeroSelectorVm列表元素间交互逻辑

[HotsBpHelper] 团队联赛,聊天框遮挡,设置/任务/好友/皮肤/技能/天赋遮挡

[HotsBpHelper] 优化相似度比较算法

[HotsBpHelper] 优化框体识别

[HotsBpHelper] 修正提示文字

[HotsBpViewModel] WPF优化
1. 加入IToastService用于消息框
2. 引入通用的ViewModelFactory,之后若加入新的ViewModel,只需在GenericViewModelFactory里声明该Factory类和CreateViewModel方法即可,无需在BootStrapper和对应构造函数中注册.

[HotsBpHelper]  不再需要IViewModelFactory

[HotsBpHelper] 优化识别体验
1. 对接临时文件,在游戏时不再扫描BP界面
2. 在OCR库无法使用时弹出提示

[HotsBpHelper] 修复提示框/任务栏框问题

[HotsBpHeper] 加入检测训练库文件

[HotsBpHelper] UI&更新优化
1. 加入tessdata检测/下载逻辑 (未测试)
2. 解决MMRView和WebFileUpdateView的启动闪烁问题
3. 新增TabSwitcherFreeWindow (继承Window), 用于确保该窗口从任务栏切换中移除

[HotsBpHelper] UX优化
1. 修复在主线程等待时无法从托盘正确关闭程序的问题
2. 初始化时隐藏托盘菜单交互按钮以避免误操作,加入提示信息

TODO: 本地化

[HotsBpHelper] 服务器请求不再堵塞构造函数
UI在加载时更流畅

[HotsBpHelper] Bug fix

[HotsBpHelper] Localization

[HotsBpHelper] 支持下载百分比

[HotsBpHelper] 加入了秒退识别警告

[HotsBpHelper] 更新图标

[HotsBpHelper] 移除不需要的工程等

remove

remove2

+++ HotsBpHelper/Utils/ScreenUtil.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Drawing;
 using System.Runtime.InteropServices;
+using System.Drawing;
 
 namespace HotsBpHelper.Utils
 {
@@ -210,7 +210,7 @@ namespace HotsBpHelper.Utils
         public static Point ToPixelPoint(this Point unitPoint)
         {
             int pixelX, pixelY;
-            TransformToPixels(unitPoint.X, unitPoint.Y, out pixelX, out pixelY);
+            TransformToPixels((int)unitPoint.X, (int)unitPoint.Y, out pixelX, out pixelY);
             return new Point(pixelX, pixelY);
         }
 
@@ -222,7 +222,7 @@ namespace HotsBpHelper.Utils
         /// <param name="pixelY">returns the Y value in pixels</param>
         /// <param name="unitX">a device independent unit value X</param>
         /// <param name="unitY">a device independent unit value Y</param>
-        public static void TransformFromPixels(int pixelX, int pixelY, out int unitX, out int unitY)
+        public static void TransformFromPixels(int pixelX, int pixelY, out double unitX, out double unitY)
         {
             IntPtr hDc = GetDC(IntPtr.Zero);
             if (hDc != IntPtr.Zero)
@@ -232,8 +232,8 @@ namespace HotsBpHelper.Utils
 
                 ReleaseDC(IntPtr.Zero, hDc);
 
-                unitX = (int) (pixelX * 96 / (double)dpiX);
-                unitY = (int) (pixelY * 96 / (double)dpiY);
+                unitX = pixelX * 96 / (double)dpiX;
+                unitY = pixelY * 96 / (double)dpiY;
             }
             else
                 throw new ArgumentNullException("Failed to get DC.");
@@ -241,16 +241,16 @@ namespace HotsBpHelper.Utils
 
         public static Point ToUnitPoint(this Point pixelPoint)
         {
-            int unitX, unitY;
-            TransformFromPixels(pixelPoint.X, pixelPoint.Y, out unitX, out unitY);
-            return new Point(unitX, unitY);
+            double unitX, unitY;
+            TransformFromPixels((int)pixelPoint.X, (int)pixelPoint.Y, out unitX, out unitY);
+            return new Point((int)unitX, (int)unitY);
 
         }
         public static Size ToUnitSize(this Size pixelSize)
         {
-            int unitWidth, unitHeight;
-            TransformFromPixels(pixelSize.Width, pixelSize.Height, out unitWidth, out unitHeight);
-            return new Size(unitWidth, unitHeight);
+            double unitWidth, unitHeight;
+            TransformFromPixels((int)pixelSize.Width, (int)pixelSize.Height, out unitWidth, out unitHeight);
+            return new Size((int)unitWidth, (int)unitHeight);
 
         }
